fix(tests): split test_checker.py into per-module test files
Replaced the single test_checker.py with six focused files:
test_normaliser.py, test_parser.py, test_fetcher.py,
test_reporter.py, test_checker_cli.py, and test_integration.py.
All 25 tests pass across the new structure.
